restructuring
moving from an eclipse to a vs code environment, so changing the layout of the file tree to reflect this.
