+++ server/Lot.java
@@ -1,0 +1,6 @@
+package server;
+
+public interface Lot {
+	TimeSlot getTimeSlot(String timekey);
+	void cullSlots(String currentTime);
+}
