+++ server/MappedLot.java
@@ -1,0 +1,5 @@
+package server;
+
+public class MappedLot implements Lot{
+
+}
