+++ server/NoOpenSpotsException.java
@@ -1,0 +1,5 @@
+package server;
+
+public class NoOpenSpotsException extends Exception{
+	public NoOpenSpotsException() {}
+}
