+++ server/Reservation.java
@@ -1,0 +1,5 @@
+package server;
+
+public class Reservation {
+	private User reserver;
+}
