+++ server/TimeSlot.java
@@ -1,0 +1,6 @@
+package server;
+
+public interface TimeSlot {
+//time slots require the key formatting ~> YYYY,MM,DD,HH
+//MM, DD, and HH are formatted in 24 hr. with a leading 0 if needed
+}
