+++ server/UnmappedTimeSlot.java
@@ -1,0 +1,5 @@
+package server;
+
+public class UnmappedTimeSlot implements TimeSlot{
+
+}
