--- src/server/Lot.java
@@ -1,6 +1,0 @@
-package server;
-
-public interface Lot {
-	TimeSlot getTimeSlot(String timekey);
-	void cullSlots(String currentTime);
-}
