--- src/server/MappedLot.java
@@ -1,5 +1,0 @@
-package server;
-
-public class MappedLot implements Lot{
-
-}
