--- src/server/NoOpenSpotsException.java
@@ -1,5 +1,0 @@
-package server;
-
-public class NoOpenSpotsException extends Exception{
-	public NoOpenSpotsException() {}
-}
