--- src/server/Reservation.java
@@ -1,5 +1,0 @@
-package server;
-
-public class Reservation {
-	private User reserver;
-}
