--- src/server/UnmappedTimeSlot.java
@@ -1,5 +1,0 @@
-package server;
-
-public class UnmappedTimeSlot implements TimeSlot{
-
-}
